Add files via upload
The main report in this folder regarding prediction of bacterial promoters from DNA sequence fragments is "Promoters-NN&ML.ipynb". The .png files refer to output of this Jupyter Notebook that does not appear on Github. The numbers in the filenames refer to the sample number from the test data from which these outputs came. The Shap image files are individual outputs from Shap analyses, and the remaining two files ("Keras_13.png" and "LR_13.png") are screenshots of portions of LIME analysis outputs.